pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proyecto-minero"
version = "1.0.0"
description = "Pipeline ETL y Soft-Sensor GP para proceso de flotación minera"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Juan Galaz", email = "juan.galaz@example.com"}
]
requires-python = ">=3.10"

keywords = [
    "mining",
    "flotation",
    "soft-sensor",
    "gaussian-process",
    "etl",
    "industrial"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "optuna>=3.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "rich>=13.0.0",
    "joblib>=1.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
mining-pipeline = "core.pipeline:main"
mining-gp = "core.models.mining_gp_pro:main"
mining-scan = "tools.scan_schema:main"

[project.urls]
Repository = "https://github.com/juan-galaz/proyecto-minero"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "config*", "tools*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.yaml", "*.yml"]

# === Configuración de herramientas de desarrollo ===

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | data
    | results
    | models
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports OK in __init__

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "data/",
    "results/",
    "models/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
